requirements.txt
discord.py
pyTableMaker==2.0.0
psutil
aiosqlite
python-dotenv
invoke
pybind11
python-config

jishaku
duckduckgo3
mcstatus
